requirements.txt
DateTime==4.3
h5py==3.6.0
joblib==1.1.0
nibabel==3.2.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.0
zope.interface==5.4.0
